+++ box/Model/GridVisorTake/MGridVisorTakeEnergy20.swift
@@ -2,7 +2,7 @@ import Foundation
 
 class MGridVisorTakeEnergy20:MGridVisorTakeEnergy
 {
-    private let kEnergyAmount:Int = 20
+    private let kEnergyAmount:Int16 = 20
     
     init()
     {
