+++ box/Model/GridVisorTake/MGridVisorTakeEnergy25.swift
@@ -2,7 +2,7 @@ import Foundation
 
 class MGridVisorTakeEnergy25:MGridVisorTakeEnergy
 {
-    private let kEnergyAmount:Int = 25
+    private let kEnergyAmount:Int16 = 25
     
     init()
     {
