+++ box/Model/GridVisorTake/MGridVisorTakeEnergy35.swift
@@ -2,7 +2,7 @@ import Foundation
 
 class MGridVisorTakeEnergy35:MGridVisorTakeEnergy
 {
-    private let kEnergyAmount:Int = 35
+    private let kEnergyAmount:Int16 = 35
     
     init()
     {
