+++ box/Model/GridVisorTake/MGridVisorTakeEnergy50.swift
@@ -2,7 +2,7 @@ import Foundation
 
 class MGridVisorTakeEnergy50:MGridVisorTakeEnergy
 {
-    private let kEnergyAmount:Int = 50
+    private let kEnergyAmount:Int16 = 50
     
     init()
     {
